backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.0
celery[redis]>=5.4.0
redis>=5.0.0
python-multipart>=0.0.12

# Dev / test
httpx>=0.27.0
pytest>=8.0.0
